+++ src/traits/search_traits.rs
@@ -4,7 +4,7 @@ use crate::commands::search::{Field, SortBy};
 pub trait Searchable {
     fn matches_text(&self, query: &str) -> bool;
     fn matches_field(&self, field: &Field, value: &str) -> bool;
-    fn matches_id(&self, id: u32) -> bool;
+    fn matches_id(&self, id: i64) -> bool;
     fn matches_tag(&self, tag: &str) -> bool;
 }
